New ForumAPI adjustments in order to keep backwards compatibility

+++ src/main/java/dev/wms/pwrapi/api/ForumAPI.java
@@ -49,7 +49,7 @@ public class ForumAPI {
             Review result = forumService.getReviewById(reviewId);
             return ResponseEntity.status(HttpStatus.OK).body(result);
         }catch (EmptyResultDataAccessException e){
-            throw new ReviewNotFoundException(reviewId);
+            throw new ReviewNotFoundException((long) reviewId);
         }
     }
